requirements.txt
PyQt6>=6.6.0
pdfplumber>=0.10.0
rapidfuzz>=3.6.0
playwright>=1.41.0
openpyxl>=3.1.2
python-docx>=1.1.0
keyring>=24.3.0
Pillow>=10.2.0
